Move TestIntegrationDeleteNestedFolders
